issue-9: circular buffer decoupled from pybind11 logic
