+++ tests/cxx/pybind11_ext/numpy_circular_buffer_tests.cpp
@@ -1,4 +1,4 @@
-#include <pybind11_ext/boost/circular_buffer.hpp>
+#include <pybind11_ext/numpy_circular_buffer.hpp>
 
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
@@ -14,50 +14,13 @@ using namespace testing;
 namespace py = pybind11;
 
 
-TEST(numpy_circular_buffer, create__char__single)
-{
-    py::detail::numpy_circular_buffer<char> buffer(1);
-    ASSERT_EQ(buffer.capacity(), size_t(getpagesize()));
-}
-
-TEST(numpy_circular_buffer, create__int__single)
-{
-    py::detail::numpy_circular_buffer<int> buffer(1);
-    ASSERT_EQ(buffer.capacity(), size_t(getpagesize() / sizeof(int)));
-}
-
-TEST(numpy_circular_buffer, create__custom__not_aligned__single)
-{
-    struct test_type
-    {
-        int field_1;
-        double field_2;
-        char field_3;
-    };
-
-    ASSERT_ANY_THROW(py::detail::numpy_circular_buffer<test_type>(1));
-}
-
-TEST(numpy_circular_buffer, create__custom__aligned__single)
-{
-    struct alignas(16) test_type
-    {
-        int field_1;
-        double field_2;
-        char field_3;
-    };
-
-    py::detail::numpy_circular_buffer<test_type> buffer(1);
-    ASSERT_EQ(buffer.capacity(), size_t(getpagesize() / sizeof(test_type)));
-}
-
 TEST(numpy_circular_buffer, to_array__int)
 {
     py::detail::numpy_circular_buffer<int> buffer(1024);
 
     // empty
     {
-        auto array = buffer.to_array();
+        auto array = py::to_array(buffer);
         ASSERT_EQ(array.size(), ssize_t(buffer.size()));
     }
 
@@ -67,7 +30,7 @@ TEST(numpy_circular_buffer, to_array__int)
         buffer.push_back(2);
         buffer.push_back(3);
 
-        auto array = buffer.to_array();
+        auto array = py::to_array(buffer);
         ASSERT_EQ(array.size(), ssize_t(buffer.size()));
         ASSERT_EQ(array.at(0), 1);
         ASSERT_EQ(array.at(1), 2);
@@ -78,7 +41,7 @@ TEST(numpy_circular_buffer, to_array__int)
     {
         buffer.pop_front();
 
-        auto array = buffer.to_array();
+        auto array = py::to_array(buffer);
         ASSERT_EQ(array.size(), ssize_t(buffer.size()));
         ASSERT_EQ(array.at(0), 2);
         ASSERT_EQ(array.at(1), 3);
@@ -89,7 +52,7 @@ TEST(numpy_circular_buffer, to_array__int)
         for (auto i = 0u; i < 2 * buffer.capacity(); i++)
             buffer.push_back(i);
 
-        auto array = buffer.to_array();
+        auto array = py::to_array(buffer);
         ASSERT_EQ(array.size(), ssize_t(buffer.size()));
 
         for (auto i = 0u; i < buffer.size(); i++)
@@ -101,7 +64,7 @@ TEST(numpy_circular_buffer, to_array__int)
         while (!buffer.empty())
             buffer.pop_front();
 
-        auto array = buffer.to_array();
+        auto array = py::to_array(buffer);
         ASSERT_EQ(array.size(), 0u);
     }
 }
@@ -120,7 +83,7 @@ TEST(numpy_circular_buffer, to_array__custom)
 
     // empty
     {
-        auto array = buffer.to_array<int>(offsetof(test_type, field_1));
+        auto array = py::to_array<test_type, int>(buffer, offsetof(test_type, field_1));
         ASSERT_EQ(array.size(), ssize_t(buffer.size()));
     }
 
@@ -130,19 +93,19 @@ TEST(numpy_circular_buffer, to_array__custom)
         buffer.push_back({2, 2.2, 'b'});
         buffer.push_back({3, 3.3, 'c'});
 
-        auto array_field_1 = buffer.to_array<int>(offsetof(test_type, field_1));
+        auto array_field_1 = py::to_array<test_type, int>(buffer, offsetof(test_type, field_1));
         ASSERT_EQ(array_field_1.size(), ssize_t(buffer.size()));
         ASSERT_EQ(array_field_1.at(0), 1);
         ASSERT_EQ(array_field_1.at(1), 2);
         ASSERT_EQ(array_field_1.at(2), 3);
 
-        auto array_field_2 = buffer.to_array<double>(offsetof(test_type, field_2));
+        auto array_field_2 = py::to_array<test_type, double>(buffer, offsetof(test_type, field_2));
         ASSERT_DOUBLE_EQ(array_field_2.size(), buffer.size());
         ASSERT_DOUBLE_EQ(array_field_2.at(0), 1.1);
         ASSERT_DOUBLE_EQ(array_field_2.at(1), 2.2);
         ASSERT_DOUBLE_EQ(array_field_2.at(2), 3.3);
 
-        auto array_field_3 = buffer.to_array<char>(offsetof(test_type, field_3));
+        auto array_field_3 = py::to_array<test_type, char>(buffer, offsetof(test_type, field_3));
         ASSERT_EQ(array_field_3.size(), ssize_t(buffer.size()));
         ASSERT_EQ(array_field_3.at(0), 'a');
         ASSERT_EQ(array_field_3.at(1), 'b');
@@ -154,7 +117,7 @@ TEST(numpy_circular_buffer, to_array__custom)
         while (!buffer.empty())
             buffer.pop_front();
 
-        auto array = buffer.to_array<int>();
+        auto array = py::to_array<test_type, int>(buffer);
         ASSERT_EQ(array.size(), 0u);
     }
 }
